# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Toilet Map API",
    "version": "1.0.0",
    "description": "Programmatic access to Toilet Map loos and administrative areas."
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Coordinates": {
        "type": "object",
        "description": "Geographic coordinates (WGS84).",
        "properties": {
          "lat": {
            "type": "number",
            "example": 51.5074
          },
          "lng": {
            "type": "number",
            "example": -0.1278
          }
        },
        "required": [
          "lat",
          "lng"
        ]
      },
      "AdminArea": {
        "type": "object",
        "description": "Administrative subdivision associated with a loo.",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Route not found"
          }
        }
      },
      "ValidationErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          {
            "type": "object",
            "properties": {
              "issues": {}
            }
          }
        ]
      },
...
      "JsonValue": {
        "description": "Arbitrary JSON value.",
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "null"
          }
        ]
      },
      "LooMutation": {
        "type": "object",
        "description": "Attributes accepted when creating or updating a loo.",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          },
          "areaId": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 24,
            "minLength": 24,
            "pattern": "^[a-f0-9]{24}$"
          },
          "accessible": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "allGender": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "attended": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "automatic": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "babyChange": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "children": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "men": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "women": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "urinalOnly": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "radar": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "noPayment": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "paymentDetails": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "removalReason": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "openingTimes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ],
            "description": "Arbitrary JSON payload describing opening times."
          },
          "location": {
            "description": "Optional location update. Null clears an existing location.",
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "CreateLooRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/LooMutation"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "minLength": 24,
                "maxLength": 24,
                "pattern": "^[a-f0-9]{24}$",
                "example": "0123456789abcdef01234567"
              }
            }
          }
        ]
      },
...
      "Loo": {
        "type": "object",
        "required": [
          "id",
          "area",
          "reports"
        ],
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "pattern": "^[a-f0-9]{24}$",
            "example": "0123456789abcdef01234567"
          },
          "geohash": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "area": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminArea"
            },
            "description": "Administrative areas linked to the loo."
          },
          "createdAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "updatedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "verifiedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "accessible": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "allGender": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "attended": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "automatic": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "babyChange": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "children": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "men": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "women": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "urinalOnly": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "noPayment": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "paymentDetails": {
            "type": [
              "string",
              "null"
            ]
          },
          "removalReason": {
            "type": [
              "string",
              "null"
            ]
          },
          "radar": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "openingTimes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opening hours metadata if known."
          },
          "reports": {
            "type": "array",
            "description": "Recent reports are populated on specific endpoints.",
            "items": {}
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "LooListResponse": {
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Loo"
            }
          },
          "count": {
            "type": "number",
            "example": 2
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/loos": {
      "get": {
        "tags": [
          "Loos"
        ],
        "summary": "Fetch loos by ID",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "description": "Comma separated IDs (?ids=a,b) or repeat the parameter (?ids=a&ids=b).",
            "schema": {
              "oneOf": [
                {
                  "type": "string",
                  "pattern": "^([a-f0-9]{24})(,[a-f0-9]{24})*$",
                  "example": "0123456789abcdef01234567,89abcdef0123456701234567"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minLength": 24,
                    "maxLength": 24,
                    "pattern": "^[a-f0-9]{24}$"
                  },
                  "example": [
                    "0123456789abcdef01234567",
                    "89abcdef0123456701234567"
                  ]
                }
              ]
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "Matching loos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LooListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Loos"
        ],
        "summary": "Create a loo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Attributes to set for the new loo.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLooRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Loo created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "A loo with the supplied ID already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}